requirements.txt
alembic==1.3.2
asn1crypto==1.2.0
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
future==0.18.2
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-telegram-bot==12.2.0
six==1.13.0
SQLAlchemy==1.3.12
tornado==6.0.3
requests==2.11.1
